Added a general purpose Tcl script -> constant C string encoder.
Added a complete zipkit implementation. (We no longer need to go out to odie for 
hard coded C sources.)

If zipfs.c is present in the Tcl sources, that is used. If zipfs.c is not present,
odie is automatically added as a sub-project and the zipfs implementation from
there is used. (Helpful while we transition from 8.6 -> 8.7)

The "main" file produced for a zipkit now includes a complete TCL_LOCAL_APPINIT
and TCL_LOCAL_MAINHOOK implementation, as well as a package index and zipfs 
mounter. This file is generated at implementation time.

Re-arranged the kit building process such that dynamically added sources have
a chance to influence compile flags